pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-google-genai>=4.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "langdetect>=1.0.9",
    "numpy>=2.4.0",
    "pdfminer-six>=20251230",
    "pi-heif>=1.1.1",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "transformers>=4.57.3",
    "unstructured>=0.18.24",
    "unstructured-inference>=0.2.4",
    "youtube-transcript-api>=1.2.3",
]
